requirements.txt
# requirements.txt — зависимости проекта

beautifulsoup4==4.12.3
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
slack-bolt==1.19.0
